+++ luxai_s2/luxai_s2/map/__init__.py
@@ -1,3 +1,2 @@
 from .board import Board
 from .position import Position
-from .weather import generate_weather_schedule, apply_weather